app/requirements.txt
Django==4.1.8
pandas==2.0.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
openpyxl==3.1.2